reached terminus on integrating espeak-ng as FFI C binding as-is
I was hoping, foolishly, that the C compiler would be clever
enough to recognize that the core functions I'm using in espeak no
longer rely on libstd and it would produce for me a stripped down
library that reflected that.

Unfortunately, it's pretty clear to me at this point that I
am going to need to create a libc shim that fakes the existence
of all the functions I never use before we can get to the
point of stripping the unused functions out.

Committing this as a checkpoint for the change in strategy.

Contemplating now doing a full fork of espeak-ng into a synthesis-only
C program that requires hard-coded .h files encoding phoneme, voice,
and dictionary for a given language, which would not have any
libstd dependencies and thus have a clean integration into Xous.

Basically, espeak-ng is actually several programs in one:
- A phoneme table generator
- A text-to-phoneme translator
- A voice definition metalanguage
- A synthesizer that takes phonemes and turns them into digital samples

The first three items are defined by human-friendly text files
and probably about 70% of espeak-ng's code base deals with parsing
the files and mapping them into static data structures that are stored
on disk. These data structures are then read in and lightly parsed
and/or hash-tabled, and then used by the speech engine to translate
text to speech.

THe current temptation is to try to knock off that very last chunk
from espeak-ng and turn it into a couple big-ish C files that are
then compiled against some .h files that encapsulate the phonemes,
text-to-phoneme dictionaries, and the voice definitions, and "vendor"
in the synthesis engine into a compact and manageable chunk that
can be optionally linked into Xous for interop with the blind
community.

This is probably a big project but I think if it's done right it
sets the text-to-speech engine on a path for durable integration into
Xous instead of just being a one-time demo.